population/requirements.txt
Flask
Flask-PyMongo
Flask-Session
dash
geopandas
plotly
gunicorn
python-dotenv
flask_cors